\name{ex-package}
\alias{ex-package}
\alias{ex}
\docType{package}
\title{
\packageTitle{ex}
}
\description{
\packageDescription{ex}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ex}
\packageIndices{ex}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ex}

Maintainer: \packageMaintainer{ex}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
